pyproject.toml
[project]
name = "rag-gateway"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "httpx>=0.27",
  "qdrant-client>=1.9",
  "tantivy>=0.22",
  "beautifulsoup4>=4.12",
  "lxml>=5.2",
]

[tool.uvicorn]
factory = false

